refactor: split out module code
* move remaining bastions to `upath`
* use `options.alias` rather than extending webpack build
